Refactoring news to news-service
