+++ back/news-service/.gitattributes
