+++ back/news-service/src/main/java/com/ds/news_service/NewsApplication.java
@@ -1,0 +1,13 @@
+package com.ds.news_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NewsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(NewsApplication.class, args);
+	}
+
+}
