+++ back/news-service/src/main/java/com/ds/news_service/ai/AiClient.java
@@ -1,0 +1,9 @@
+package com.ds.news_service.ai;
+
+import com.ds.news_service.models.record.AiEnrichmentResult;
+import reactor.core.publisher.Mono;
+
+public interface AiClient {
+    Mono<AiEnrichmentResult> enrich(String text);
+
+}
