+++ back/news-service/src/main/java/com/ds/news_service/controllers/AnalyticsController.java
@@ -1,0 +1,14 @@
+package com.ds.news_service.controllers;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Component
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/analytics")
+public class AnalyticsController {
+
+}
