+++ back/news-service/src/main/java/com/ds/news_service/controllers/NewsController.java
@@ -1,0 +1,23 @@
+package com.ds.news_service.controllers;
+
+import com.ds.news_service.models.payloads.RawNewsPayload;
+import com.ds.news_service.services.NewsService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.annotation.*;
+import reactor.core.publisher.Mono;
+
+@Component
+@RestController
+@RequestMapping("/news")
+@RequiredArgsConstructor
+public class NewsController {
+
+    private final NewsService newsService;
+
+    @GetMapping("/{id}")
+    public Mono<RawNewsPayload>  getNews(@PathVariable String id) {
+        return newsService.getNewsById(id);
+    }
+
+}
