+++ back/news-service/src/main/java/com/ds/news_service/models/record/AiEnrichmentResult.java
@@ -1,0 +1,8 @@
+package com.ds.news_service.models.record;
+
+public record AiEnrichmentResult(
+//        String summary,
+        float[] embedding,
+        String language
+) {
+}
