+++ back/news-service/src/test/java/com/ds/news_service/NewsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ds.news_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NewsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
