--- back/news/src/main/java/com/ds/news/NewsApplication.java
@@ -1,13 +1,0 @@
-package com.ds.news;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NewsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(NewsApplication.class, args);
-	}
-
-}
