--- back/news/src/main/java/com/ds/news/controllers/AnalyticsController.java
@@ -1,14 +1,0 @@
-package com.ds.news.controllers;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Component
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/analytics")
-public class AnalyticsController {
-
-}
