--- back/news/src/main/java/com/ds/news/models/record/AiEnrichmentResult.java
@@ -1,8 +1,0 @@
-package com.ds.news.models.record;
-
-public record AiEnrichmentResult(
-//        String summary,
-        float[] embedding,
-        String language
-) {
-}
