--- back/news/src/main/java/com/ds/news/repositories/NewsRepository.java
@@ -1,11 +1,0 @@
-package com.ds.news.repositories;
-
-import com.ds.news.models.payloads.RawNewsPayload;
-import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
-import org.springframework.stereotype.Repository;
-import reactor.core.publisher.Mono;
-
-@Repository
-public interface NewsRepository extends ReactiveMongoRepository<RawNewsPayload, String> {
-    Mono<RawNewsPayload> findByExternalId(String externalId);
-}
