--- back/news/src/main/java/com/ds/news/services/NewsPersistenceService.java
@@ -1,90 +1,0 @@
-package com.ds.news.services;
-
-import com.ds.news.ai.AiClient;
-import com.ds.news.models.payloads.RawNewsPayload;
-import com.ds.news.repositories.NewsRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import reactor.core.publisher.Mono;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class NewsPersistenceService {
-    private final NewsRepository newsRepository;
-    private final AiClient aiClient;
-
-    public Mono<RawNewsPayload> persistAndEnrich(RawNewsPayload payload) {
-        String semanticText = buildNewsSemanticText(payload);
-
-        return aiClient.enrich(semanticText)
-                .map(result -> {
-                    payload.setAiSummary(payload.getDescription());
-                    payload.setEmbedding(result.embedding());
-                    payload.setLanguage(result.language());
-                    return payload;
-                })
-                .flatMap(newsRepository::save)
-                .doOnSuccess(saved ->
-                        log.info("News enriched and persisted {}", saved.getId())
-                );
-    }
-
-    private String detectLanguage(String content) {
-        if (content == null) return "unknown";
-        return "en";
-    }
-
-    private String buildNewsSemanticText(RawNewsPayload news) {
-        return """
-        Title: %s
-        Description: %s
-        Category: %s
-        Author: %s
-        Content: %s
-        """.formatted(
-                    safe(news.getTitle()),
-                    safe(news.getDescription()),
-                    safe(news.getCategory()),
-                    safe(news.getAuthor()),
-                    safe(news.getContent())
-            );
-    }
-
-    private String safe(String value) {
-        return value == null ? "" : value;
-    }
-}
-
-/*
-*  TODO: for user service
-* private String buildUserPreferenceText(UserRecommendationProfile user) {
-    return """
-    Interested in: %s
-    """.formatted(
-        String.join(", ", user.getPreferences().getPreferredCategories())
-    );
-*@Service
-@RequiredArgsConstructor
-public class UserEmbeddingService {
-
-    private final AiClient aiClient;
-    private final UserRepository userRepository;
-
-    public Mono<UserRecommendationProfile> updateEmbedding(
-            UserRecommendationProfile user) {
-
-        String text = buildUserPreferenceText(user);
-
-        return aiClient.enrich(text)
-            .map(result -> {
-                user.setPreferenceEmbedding(result.embedding());
-                user.setEmbeddingUpdatedAt(Instant.now());
-                return user;
-            })
-            .flatMap(userRepository::save);
-    }
-}
-
-}*/
