--- back/news/src/main/resources/application.yaml
@@ -1,37 +1,0 @@
-spring:
-  application:
-    name: news-service
-  kafka:
-    bootstrap-servers: ${KAFKA_SERVER:http://localhost:29092}
-    producer:
-      key-serializer: org.apache.kafka.common.serialization.StringSerializer
-      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
-      acks: all
-      retries: 3
-    consumer:
-      group-id: raw-news-consumer-group
-      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
-      auto-offset-reset: earliest
-      properties:
-        spring.json.trusted.packages: "*"
-        spring.json.value.default.type: com.ds.news.models.payloads.RawNewsPayload
-        spring.json.use.type.headers: false
-  mongodb:
-    uri: ${MONGODB_URL:mongodb://newsadmin:newspass@localhost:5060/news?authSource=admin}
-
-news:
-  topic: raw-news
-  enriched_news_topic: enriched_news
-
-ai:
-  embedding_uri: ${EMBEDDING_URI:http//localhost:8999}
-
-server:
-  port: 8902
-eureka:
-  instance:
-    hostname: ${EUREKA_HOST:localhost}
-  client:
-    service-url:
-      defaultZone: ${DEFAULT_ZONE:http://localhost:8901/eureka}
