--- back/news/src/test/java/com/ds/news/NewsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ds.news;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NewsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
